# perf: Rust improvements